fixed a bug with time and date pickers, added jsdocs to utils and minor changes and improvements
