+++ src/components/activity-edit-form/state/activityForm.enums.ts
@@ -1,0 +1,25 @@
+export enum ActionType {
+    SET_ICON_KEY = 'SET_ICON_KEY',
+    SET_TITLE = 'SET_TITLE',
+    SET_DESCRIPTION = 'SET_DESCRIPTION',
+    SET_CATEGORIES = 'SET_CATEGORIES',
+    SET_RATING = 'SET_RATING',
+    SET_START_TIME = 'SET_START_TIME',
+    SET_END_TIME = 'SET_END_TIME',
+    SET_ALERT = 'SET_ALERT',
+    CLEAR_ALERT = 'CLEAR_ALERT',
+    SET_IS_NOW = 'SET_IS_NOW',
+    SET_TIME = 'SET_TIME',
+    RESET = 'RESET',
+    SET_DELETE_GUARD = 'SET_DELETE_GUARD',
+}
+export enum Alerts {
+    ICON_KEY = 'iconKey',
+    TITLE = 'title',
+    DESCRIPTION = 'description',
+    CATEGORIES = 'categoryIds',
+    RATING = 'rating',
+    START_TIME = 'startTime',
+    END_TIME = 'endTime',
+    DELETE_GUARD = 'deleteGuard',
+}
