+++ src/styles/css-variables.scss
@@ -2,12 +2,12 @@
 @use 'css-variables/sizes';
 @use 'css-variables/colors';
 @use 'css-variables/animations';
-@use 'css-variables/general';
+@use 'css-variables/dimensions';
 
 :root {
     @include fonts.fonts;
     @include sizes.sizes;
     @include colors.colors;
     @include animations.animations;
-    @include general.general;
+    @include dimensions.dimensions;
 }
