+++ src/styles/css-variables/_colors.scss
@@ -5,6 +5,11 @@
     --color-primary: #4bb3d9;
     --color-secondary: #80cfa9;
 
+    /* Scrollbar */
+    --scrollbar-thumb-color: #dfdfdf;
+    --scrollbar-thumb-color-hover: #c7c7c7;
+    --scrollbar-track-color: #ffffff;
+
     /* Additional Colors */
     --color-heart: #eb4f4f;
     --color-count: #ef8427;
