+++ src/utils/time.ts
@@ -4,8 +4,31 @@ import {
     eachDayOfInterval,
     addWeeks,
     format,
+    set,
 } from 'date-fns';
 
+/**
+ * Copies the year, month, and date from one Date object to another.
+ *
+ * @param {Date} targetDate The Date object to modify.
+ * @param {Date} sourceDate The Date object to copy from.
+ * @returns {Date} The targetDate with the year, month, and date copied from the sourceDate.
+ */
+export const copyDate = (targetDate: Date, sourceDate: Date): Date =>
+    set(targetDate, {
+        year: sourceDate.getFullYear(),
+        month: sourceDate.getMonth(),
+        date: sourceDate.getDate(),
+    });
+
+/**
+ * Formats a time range as a string.
+ *
+ * @param {Date} startTime The start time of the range.
+ * @param {Date} [endTime] The end time of the range. If not provided, only the start time is formatted.
+ * @param {boolean} [isNow] If true, the function immediately returns the string 'now'.
+ * @returns {string} The formatted time range.
+ */
 export const formatTimeRange = (
     startTime: Date,
     endTime?: Date,
@@ -24,6 +47,14 @@ export const formatTimeRange = (
     return formattedStartTime;
 };
 
+/**
+ * Generates an array of Date objects for each day of the week, considering the week that contains the specified date.
+ * The function determines the start of the week for the given date, then applies the optional offset in weeks from that start.
+ *
+ * @param {Date} date The date within the week to be considered.
+ * @param {number} [offset] The number of weeks to offset from the start of the week of the given date.
+ * @returns {Date[]} An array of Date objects representing each day of the week.
+ */
 export const getArrayOfWeekDatesFromDate = (
     date: Date,
     offset?: number
